+++ Source/MQTTnet/Server/Internal/MqttClientSessionsManager.cs
@@ -795,11 +795,9 @@ namespace MQTTnet.Server
 
         async Task<ValidatingConnectionEventArgs> ValidateConnection(MqttConnectPacket connectPacket, IMqttChannelAdapter channelAdapter)
         {
-            var eventArgs = new ValidatingConnectionEventArgs(connectPacket, channelAdapter)
-            {
-                SessionItems = new ConcurrentDictionary<object, object>()
-            };
-
+            // TODO: Load session items from persisted sessions in the future.
+            var sessionItems = new ConcurrentDictionary<object, object>();
+            var eventArgs = new ValidatingConnectionEventArgs(connectPacket, channelAdapter, sessionItems);
             await _eventContainer.ValidatingConnectionEvent.InvokeAsync(eventArgs).ConfigureAwait(false);
 
             // Check the client ID and set a random one if supported.