+++ Source/MQTTnet/Server/Internal/MqttClientStatistics.cs
@@ -10,14 +10,14 @@ namespace MQTTnet.Server
 {
     public sealed class MqttClientStatistics
     {
-        long _receivedApplicationMessagesCount;
+        Statistics _statistics = new Statistics     // mutable struct, don't make readonly!
+        {
+            // Start with 1 because the CONNACK packet is not counted here.
+            _receivedPacketsCount = 1,
 
-        // Start with 1 because the CONNACK packet is not counted here.
-        long _receivedPacketsCount = 1;
-        long _sentApplicationMessagesCount;
-
-        // Start with 1 because the CONNECT packet is not counted here.
-        long _sentPacketsCount = 1;
+            // Start with 1 because the CONNECT packet is not counted here.
+            _sentPacketsCount = 1
+        };
 
         public MqttClientStatistics()
         {
@@ -30,7 +30,6 @@ namespace MQTTnet.Server
         }
 
         public DateTime ConnectedTimestamp { get; }
-
         public DateTime LastNonKeepAlivePacketReceivedTimestamp { get; private set; }
 
         /// <summary>
@@ -43,13 +42,13 @@ namespace MQTTnet.Server
         /// </summary>
         public DateTime LastPacketSentTimestamp { get; private set; }
 
-        public long ReceivedApplicationMessagesCount => Interlocked.Read(ref _receivedApplicationMessagesCount);
+        public long SentApplicationMessagesCount => Volatile.Read(ref _statistics._sentApplicationMessagesCount);
 
-        public long ReceivedPacketsCount => Interlocked.Read(ref _receivedPacketsCount);
+        public long ReceivedApplicationMessagesCount => Volatile.Read(ref _statistics._receivedApplicationMessagesCount);
 
-        public long SentApplicationMessagesCount => Interlocked.Read(ref _sentApplicationMessagesCount);
+        public long SentPacketsCount => Volatile.Read(ref _statistics._sentPacketsCount);
 
-        public long SentPacketsCount => Interlocked.Read(ref _sentPacketsCount);
+        public long ReceivedPacketsCount => Volatile.Read(ref _statistics._receivedPacketsCount);
 
         public void HandleReceivedPacket(MqttPacket packet)
         {
@@ -61,11 +60,11 @@ namespace MQTTnet.Server
             // This class is tracking all values from Clients perspective!
             LastPacketSentTimestamp = DateTime.UtcNow;
 
-            Interlocked.Increment(ref _sentPacketsCount);
+            Interlocked.Increment(ref _statistics._sentPacketsCount);
 
             if (packet is MqttPublishPacket)
             {
-                Interlocked.Increment(ref _sentApplicationMessagesCount);
+                Interlocked.Increment(ref _statistics._sentApplicationMessagesCount);
             }
 
             if (!(packet is MqttPingReqPacket || packet is MqttPingRespPacket))
@@ -73,6 +72,7 @@ namespace MQTTnet.Server
                 LastNonKeepAlivePacketReceivedTimestamp = LastPacketReceivedTimestamp;
             }
         }
+        public void ResetStatistics() => _statistics.Reset();
 
         public void HandleSentPacket(MqttPacket packet)
         {
@@ -84,20 +84,29 @@ namespace MQTTnet.Server
             // This class is tracking all values from Clients perspective!
             LastPacketReceivedTimestamp = DateTime.UtcNow;
 
-            Interlocked.Increment(ref _receivedPacketsCount);
+            Interlocked.Increment(ref _statistics._receivedPacketsCount);
 
             if (packet is MqttPublishPacket)
             {
-                Interlocked.Increment(ref _receivedApplicationMessagesCount);
+                Interlocked.Increment(ref _statistics._receivedApplicationMessagesCount);
             }
         }
 
-        public void ResetStatistics()
+        private struct Statistics
         {
-            Interlocked.Exchange(ref _sentApplicationMessagesCount, 0);
-            Interlocked.Exchange(ref _receivedApplicationMessagesCount, 0);
-            Interlocked.Exchange(ref _sentPacketsCount, 0);
-            Interlocked.Exchange(ref _receivedPacketsCount, 0);
+            public long _receivedPacketsCount;
+            public long _sentPacketsCount;
+
+            public long _receivedApplicationMessagesCount;
+            public long _sentApplicationMessagesCount;
+
+            public void Reset()
+            {
+                Volatile.Write(ref _receivedPacketsCount, 0);
+                Volatile.Write(ref _sentPacketsCount, 0);
+                Volatile.Write(ref _receivedApplicationMessagesCount, 0);
+                Volatile.Write(ref _sentApplicationMessagesCount, 0);
+            }
         }
     }
 }