+++ Source/MQTTnet/Server/MqttServer.cs
@@ -28,8 +28,6 @@ namespace MQTTnet.Server
         readonly MqttRetainedMessagesManager _retainedMessagesManager;
         readonly IMqttNetLogger _rootLogger;
 
-        readonly IDictionary _sessionItems = new ConcurrentDictionary<object, object>();
-
         CancellationTokenSource _cancellationTokenSource;
 
         public MqttServer(MqttServerOptions options, IEnumerable<IMqttServerAdapter> adapters, IMqttNetLogger logger)
@@ -169,8 +167,11 @@ namespace MQTTnet.Server
 
         MqttClientSessionsManager IMqttServerExtensibility.MqttClientSessionsManager => _clientSessionsManager;
 
-        IDictionary IMqttServerExtensibility.SessionItems => _sessionItems;
-
+        /// <summary>
+        ///     Gives access to the session items which belong to this server. This session items are passed
+        ///     to several events instead of the client session items if the event is caused by the server instead of a client.
+        /// </summary>
+        public IDictionary ServerSessionItems { get; } = new ConcurrentDictionary<object, object>();
 
         public Task DeleteRetainedMessagesAsync()
         {
@@ -233,9 +234,11 @@ namespace MQTTnet.Server
                 throw new NotSupportedException("Injected application messages must contain a topic. Topic alias is not supported.");
             }
 
+            var sessionItems = injectedApplicationMessage.CustomSessionItems ?? ServerSessionItems;
+            
             return _clientSessionsManager.DispatchApplicationMessage(
                 injectedApplicationMessage.SenderClientId,
-                _sessionItems,
+                sessionItems,
                 injectedApplicationMessage.ApplicationMessage,
                 cancellationToken);
         }